% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_panel.R
\name{balance_panel}
\alias{balance_panel}
\title{Make panel data balanced}
\usage{
balance_panel(
  df,
  type = c("fill_NA", "drop_individuals", "drop_times"),
  index = NULL,
  individual_var = NULL,
  time_seq = NULL
)
}
\arguments{
\item{df}{an object of class `data.frame`, `tibble`}

\item{type}{character, one of `"fill_NA"`,
`"drop_individuals"`, or `"drop_times"`, see
**Details**,}

\item{index}{only relevant for `data.frame` interface; if
`NULL`, the first two columns of the data.frame are
assumed to be the index variables; if not `NULL`, both
dimensions ('individual', 'time') need to be specified by
`index` as character of length 2 for data frame.}

\item{individual_var}{default is NULL; time invariant variables in the dataset, when balancing dataset they would keep the same by `"updown"` search.}

\item{time_seq}{default is NULL; window for balancing panel data set.}
}
\value{

}
\description{
Make panel data balanced
}
\examples{
data("EmplUK", package = "plm")
EmplUK \%>\%
select(year, firm) \%>\%
filter(firm \%in\% c(1:10)) \%>\%
table()

balance_panel(EmplUK, type = "fill_NA", individual_var = c("sector"))
}
