% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{download_file}
\alias{download_file}
\title{Download file}
\usage{
download_file(link, dfile)
}
\arguments{
\item{link}{the url}

\item{dfile}{the destination file}
}
\value{

}
\description{
Download file
}
\examples{
download_file("https://www.sec.gov/Archives/edgar/full-index/form.idx", "form.txt")
}
