% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stata.R
\name{find_stata}
\alias{find_stata}
\title{Find Stata}
\usage{
find_stata(message = TRUE)
}
\arguments{
\item{message}{}
}
\value{
The path of STATA
}
\description{
This function is from Doug Hemken, Statamarkdown package
}
