% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{fred}
\alias{fred}
\title{FRED}
\usage{
fred(Dataset = character(), start.date, end.date)
}
\arguments{
\item{Dataset}{The data set want to use in FRED}

\item{start.date}{The start date of the data set}

\item{end.date}{The end date of the data set}
}
\value{

}
\description{
FRED
}
\examples{
fred("FEDFUNDS", "1954-01-01", "2021-08-01")
}
