% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_reg.R
\name{lm_robust}
\alias{lm_robust}
\title{Ordinary Least Squares with Robust Standard Errors}
\usage{
lm_robust(
  formula,
  data,
  subset,
  weights,
  na.action,
  se_type = "HC1",
  clusters = NULL,
  model = TRUE,
  x = FALSE,
  y = FALSE,
  singular.ok = TRUE,
  contrasts = NULL,
  offset,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Ordinary Least Squares with Robust Standard Errors
}
