% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{read_stata}
\alias{read_stata}
\title{Read Stata dta file from given path or website}
\usage{
read_stata(path, webuse = FALSE)
}
\arguments{
\item{path}{The path of dta file}
}
\value{
Return is a data.frame object
}
\description{
Read Stata dta file from given path or website
}
\examples{
read_stata("hiway")
Load (Minnesota Highway Data, 1973)

read_stata("https://www.stata-press.com/data/r17/union")

read_stata("data.dta")
}
