% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_reg.R
\name{summary_lm_robust}
\alias{summary_lm_robust}
\title{Summary lm_robust object}
\usage{
summary_lm_robust(object, vcov. = vcovHC, ...)
}
\arguments{
\item{vcov.}{}
}
\value{

}
\description{
Summary lm_robust object
}
